added air quality sensor for EEP A5-09-0C
